fix: 清理编译期废弃 API 与冗余代码警告

+++ feature-home/src/main/java/app/amisles/hanime/feature/home/HomeScreen.kt
@@ -411,7 +411,7 @@ private suspend fun LazyListState.smoothScrollToItem(
         return
     }
 
-    val avgSize = visibleItems.map { it.size }.average().toDouble().toInt()
+    val avgSize = visibleItems.map { it.size }.average().toInt()
     if (avgSize <= 0) {
         animateScrollToItem(index, offset)
         return
